@@ -1,28 +1,17 @@
 {
-  "name": "drag-and-drop-trello",
-  "version": "1.0.0",
-  "description": "Drag and Drop implementation using react. Kanban board format is used.\nHTML and style credits: https://codepen.io/Paolo-Duzioni/pen/WNrJdYd ",
+  "name": "trellike",
+  "version": "0.0.1",
+  "description": "Starting point for a kanban board UI",
   "keywords": [],
   "main": "src/index.js",
   "dependencies": {
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-scripts": "4.0.0"
+    "react": "18",
+    "react-dom": "18"
   },
   "devDependencies": {
-    "@babel/runtime": "7.13.8",
-    "typescript": "4.1.3"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
-  },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
+    "dev": "esbuild --bundle ./src/index.js --outdir=public --loader:.js=jsx --servedir=public",
+    "build": "esbuild --bundle ./src/index.js --outdir=public --loader:.js=jsx --minify"
+  }
 }
