@@ -1,12 +1,12 @@
-import { StrictMode } from "react";
+import React, { StrictMode } from "react";
 import ReactDOM from "react-dom";
 
 import App from "./App";
 
 const rootElement = document.getElementById("root");
-ReactDOM.render(
+const root = ReactDOM.createRoot(rootElement);
+root.render(
   <StrictMode>
     <App />
-  </StrictMode>,
-  rootElement
+  </StrictMode>
 );
